Removed js-code of BaseTextarea
